Restructure existing notifications
